create new file and rename structs
